+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Unreleased
+  - Fix bug where previous submissions field values were bound to log lines of current submission
+  - Update packages with security issues
 
 ### 2.15.1 2019-03-14
   - Bound case_id , tx_id and user_id to logger